# Use btree to search fields in DFSchema