@alamb 2023-10-26
Related comment: https://github.com/apache/arrow-datafusion/issues/7698#issuecomment-1781787244