@oleggator 2023-10-26
> Is there a reason to use a b-tree ( O(log⁡n) ) vs a hash map ( O(1) )?

Using b-tree we can query all fields matching to a "prefix" in one O(logn) hop (`column.*.*.*`, `column.table.*.*`, `column.table.schema.*`, `column.table.schema.catalog`).
It is used in `fields_with_unqualified_name` method to query all fields by specific name.