@crepererum 2023-10-27
> > Is there a reason to use a b-tree ( O(log⁡n) ) vs a hash map ( O(1) )?
> 
> Using b-tree we can query all fields matching to a "prefix" in one O(logn) hop (`column.*.*.*`, `column.table.*.*`, `column.table.schema.*`, `column.table.schema.catalog`). It is used in `fields_with_unqualified_name` method to query all fields by specific name.

Is that such a common operation that it is worth to keep an expensive index on every single schema in the query graph? I think the planner that resolves these names can easily order the fields and build this index locally.